test: split 06_cluster_regressions into six themed suites
2243 lines / 72 tests in one file became six `06_cluster_*` suites (largest
700 lines) plus `tests/cluster/mod.rs` for the 12 shared conversion helpers,
following the existing `mod common;` / `mod helpers;` convention. Test targets
are auto-discovered, so no Cargo.toml change.

All 72 tests preserved verbatim and passing; the split is by theme —
package/engine clusters, math, bibliography, frontmatter, TOC/navigation,
standalone+subfiles.

Adds `convert_and_post_clean`, which gates the POST stage on zero errors.
`convert_and_post` only inherits `convert_to_xml`'s CORE-stage gate, and the
bibliography is built during post-processing — so a post-stage error flood
passed every bibliography guard silently (17 errors on one fixture, 203 on a
witness, both green). That is the false negative the signal-integrity rule
forbids.
